change data variable type from int to long in GetForecastDetailsUseCase.

+++ domain/src/main/java/nyc/insidout/weathervane/domain/interactor/GetForecastDetailsUseCase.java
@@ -36,13 +36,13 @@ public class GetForecastDetailsUseCase implements UseCase<GetForecastDetailsUseC
     * Class used to carry request parameters to GetForecastDetailsUseCase
     */
     public static final class RequestParam implements UseCase.RequestParam{
-        private final int date;
+        private final long date;
 
-        public RequestParam(final int date){
+        public RequestParam(final long date){
             this.date = date;
         }
 
-        public int getDate(){
+        public long getDate(){
             return date;
         }
     }
